feat: 餐饮门店

+++ database/migrations/2023_09_09_143547_create_restaurants_table.php
@@ -23,8 +23,8 @@ class CreateRestaurantsTable extends Migration
             $table->longText('food_image_list')->comment('菜品图片列表');
             $table->longText('environment_image_list')->comment('环境图片列表');
             $table->longText('price_image_list')->comment('价目表图片列表');
-            $table->float('longitude')->comment('经度');
-            $table->float('latitude')->comment('纬度');
+            $table->decimal('longitude', 9, 6)->comment('经度');
+            $table->decimal('latitude', 8, 6)->comment('纬度');
             $table->string('address')->comment('具体地址');
             $table->float('rate')->default(0)->comment('餐馆综合评分');
             $table->float('taste_rate')->default(0)->comment('餐馆口味评分');
